server/requirements.txt
pypdf
langchain
chromadb # Vector storage
pytest
boto3
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.1
python-dotenv==1.0.1
langchain==0.3.2
langchain-community==0.3.1
langchain-chroma==0.1.4
langchain-huggingface==0.1.0
langchain-google-genai==2.0.3
chromadb==0.5.11
sentence-transformers==3.1.1
pypdf==5.0.0